Completed evaluator and simple alpha beta, added simple UCI

+++ Dragonfly.Engine/Interfaces/IEvaluator.cs
@@ -7,6 +7,6 @@ namespace Dragonfly.Engine.Interfaces
 {
     public interface IEvaluator
     {
-        int Evaluate(Position position);
+        short Evaluate(Position position);
     }
 }
